backend/requirements.txt

python-dotenv==1.0.0
fastapi==0.116.1
uvicorn[standard]==0.27.1
pyswisseph==2.10.3.2
openai>=1.0.0
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
pytz==2023.3
